feat: 이미지 저장 버튼, 미리보기 모달, 미리보기 모달 컨텐츠 분리 (#118)
