+++ src/components/domain/preview/ProfilePreviewContent.tsx
@@ -1,0 +1,7 @@
+import { useUserStore } from '@/stores/userStore';
+
+export default function ProfilePreviewContent() {
+  const userId = useUserStore((state) => state.user?.userId);
+  console.log(userId);
+  return <></>;
+}
